Get rid of EventController
